issue#57: Finace block
1. Added OOTB Carousel Block.
2. I have moved the Social-Share and DiscoverMore anchors to the Auto block.
3. Added few sections.
